% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sfc_path.R
\name{geom_sfc_path}
\alias{geom_sfc_path}
\title{Draw the space-filling curve itself}
\usage{
geom_sfc_path(...)
}
\arguments{
\item{...}{Further arguments to be passed to \code{\link[ggplot2]{geom_path}}}
}
\description{
Draw the space-filling curve itself
}
