% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_points.R
\name{geom_sfc_points}
\alias{geom_sfc_points}
\title{Draw points on a space-filling curve}
\usage{
geom_sfc_points(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  ends = range(data$pos),
  curve = c("hilbert", "moore", "sierpinski", "flowsnake"),
  order = NULL,
  draw_curve = TRUE,
  curve_col = "black",
  curve_linetype = 1L,
  curve_size = 0.5,
  arrow = NULL,
  lineend = "butt",
  linejoin = "round",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
Add Description...
}
