% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsfc.R
\name{ggsfc}
\alias{ggsfc}
\title{Setup a ggsfc plot object}
\usage{
ggsfc(data, ...)
}
\arguments{
\item{...}{Further arguments to be passed to \code{\link{sfcurve}}}
}
\value{

}
\description{
Setup a ggsfc plot object
}
