% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{plot.sfcurve}
\alias{plot.sfcurve}
\title{Plot an sfcurve object}
\usage{
\method{plot}{sfcurve}(x, y, ...)
}
\arguments{
\item{x}{An \code{sfcurve} object}

\item{y}{Not used}

\item{...}{Further argument to be passed to or from other methods.}
}
\description{
Plot an sfcurve object
}
