% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{st_as_sf}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sf}{\code{\link[sf]{st_as_sf}}}
}}

