% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{sfcurve}
\alias{sfcurve}
\title{Make a space-filling curve object}
\usage{
sfcurve(
  data,
  pos_vars,
  curve = c("hilbert", "flowsnake", "sierpinski", "moore"),
  order = NULL,
  len = NULL,
  limits = NULL,
  split_by = NULL,
  split_arrange = c("interleave", "tile")
)
}
\arguments{
\item{curve}{Which space-filling curve to make? Must be one of: "hilbert", "flowsnake",
"sierpinski", or "moore". See details for descriptions.}

\item{order}{What order should the curve be? This controls the "length" of the fractal curve
(e.g. how many points are there in it?). In other words, higher order curves fill the space more.}

\item{len}{An alternative way to specify order. Instead, specify the "length" of the curve you want.
If specified, the order will be determined as the order producing a curve length closest to \code{len}}

\item{limits}{The values you want to use as the beginning and end of your curve. This will typically
be determined by your data. See details for examples.}
}
\value{
An object of class \code{sfcurve}
}
\description{
This function creates a space-filling curve of a particular order and type.
}
\examples{
sf_curve <- sfcurve("moore", order = 5)
plot(sf_curve, col = rainbow(nrow(sf_curve)))
}
