% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sfc_points.R
\name{stat_sfc_point}
\alias{stat_sfc_point}
\title{Stat to draw points on a space-filling curve}
\usage{
stat_sfc_point(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
Stat to draw points on a space-filling curve
}
